[fix] 스터디 메인 함수 이동
